+++ utility/ilog2.h
@@ -40,7 +40,7 @@ namespace HL {
   static inline int ilog2 (size_t sz)
   {
     sz = (sz << 1) - 1;
-    return (int) ((sizeof(unsigned long) * 8) - __builtin_clzl(sz) - 1);
+    return (int) ((sizeof(size_t) * 8) - __builtin_clzl(sz) - 1);
   }
 #else
   static inline int ilog2 (size_t v) {
